+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.0.3-pre
 * Change exec to display stdout and stderr in chunks
 * Add command arguments validation support
+* Display help when command is unspecified
 
 ### 0.0.2
 * Colourise success and error output