+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add command arguments validation support
 * Display help when command is unspecified
 * Display error message when command is unknown
+* Add help examples support
 
 ### 0.0.2
 * Colourise success and error output